code2DFD/requirements.txt
Flask>=2.1.2
plantuml==0.3.0
PyYAML==6.0
ruamel.base==1.0.0
PyDriller==2.6.0
lxml==5.3.0
Werkzeug>=2.0,<3.0
ruamel.yaml==0.17.21
six==1.16.0
python-dotenv>=1.0.0
flask-cors==3.0.10